+++ api/v1/dogu_types.go
@@ -88,6 +88,7 @@ const (
 	PendingHealthStatus     HealthStatus = ""
 	AvailableHealthStatus   HealthStatus = "available"
 	UnavailableHealthStatus HealthStatus = "unavailable"
+	UnknownHealthStatus     HealthStatus = "unknown"
 )
 
 // DoguStatus defines the observed state of a Dogu.
@@ -234,7 +235,7 @@ func (d *Dogu) changeRequeuePhase(ctx context.Context, client client.Client, pha
 	return d.Update(ctx, client)
 }
 
-// ChangeRequeuePhaseWithRetry refreshes the dogu resource and tries to set the requeue phase. 
+// ChangeRequeuePhaseWithRetry refreshes the dogu resource and tries to set the requeue phase.
 // If a conflict error occurs this method will retry the operation.
 func (d *Dogu) ChangeRequeuePhaseWithRetry(ctx context.Context, client client.Client, phase string) error {
 	return retry.OnConflict(func() error {